requirements.txt
selenium==4.18.1
python-dotenv==1.0.1
openai==1.12.0
pdfplumber==0.10.3
gspread
oauth2client